Jackson Changes
I added another platform with a tree and mushroom and changed the player speed a little bit. I am only curious about all of the extra files that this generated if someone wants to take a look at why it did this. I don't think we need all of them but this is what it did.