Added Animations for Shooting
Added animations for shooting. Pressing R will activate the shooting animation. If you are running the animation for shooting and running will play. The gun will also stay risen for 1 second exactly before it returns to idle.

KNOWN BUG: Once you are in an animation for either shooting stationary or shooting while running the animation will not change until the timer is up. This is KNOWN and WILL BE FIXED.