Fixed health bug
Fixed bug where health would be reset upon entering a different room.

Now health is in a separate script called GameState. Use this script for any variables you want to remain constant throughout the game and do not want to be reloaded with the character every time it is reloaded.