Adds Story Scene and Fade Transition
Adds a more "temporary" story scene while waiting for asset and more story guidance. Additionally adds a fade transition affect that can be applied anywhere. This could also be applied to doors. The timing of the animation can be updated and changed as well, pending on other team member's feedback